test(tap): create one journey for message meet
